sg-security-ai/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpclib==0.4.8
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
protobuf==5.29.5
PyJWT==2.10.1
pymongo==4.13.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
viam-sdk==0.51.0
Werkzeug==3.1.3
